requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyarrow>=12.0.0
questdb>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
urllib3>=1.26.0
certifi>=2023.0.0
